requirements.txt
streamlit
gtts
nltk
gtts
newspaper3k
langdetect
